requirements.txt
channels
djangorestframework
django-cors-headers
channels-redis

